10-agents_as_tool-app/pyproject.toml
[project]
name = "agents-as-tool-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents[litellm]>=0.1.0",
    "python-dotenv>=1.1.1",
]
